D3ao: batch 4 - damageGibAnimCurves + context-menu options
- interface.hpp damageGibAnimCurves (map<int, vector<int>>) ->
  DynamicMapI32T<DynamicArrayS32>; NEW MK_DynArrayS32 map value kind (19)
  + dynarrs32_value_free/copy in dynamic_map.odin (POD array ops)
- getContextMenuOptionsForItem/getContextTooltipOptionsForItem ->
  DynamicArrayS32 return; sweep of 10+ consumer files (drawstatus,
  playerinventory, GameUI) with enum casts at param sites; erase-iterator
  loops converted
- DynamicArrayT gained erase(T*) (returns next) + insert(T*, const T&)
  (std::vector::insert-at-begin semantics); erase(0) -> erase((int64_t)0)
  to disambiguate

94 header vectors remain (damageGib + context-menu done).

+++ src/magic/castSpell.cpp
@@ -1125,7 +1125,7 @@ Entity* castSpell(Uint32 caster_uid, spell_t* spell, bool using_magicstaff, bool
 					achievementStrobeVec[player].push_back(ticks);
 					if ( achievementStrobeVec[player].size() > 20 )
 					{
-						achievementStrobeVec[player].erase(0);
+						achievementStrobeVec[player].erase((int64_t)0);
 					}
 					Uint32 timeDiff = achievementStrobeVec[player].back() - achievementStrobeVec[player].front();
 					if ( achievementStrobeVec[player].size() == 20 )
